src/shared/parameter_v2/scripts/requirements.txt
PyYAML==5.3.1
jsonschema==2.6.0
colorama==0.3.7

